Add files via upload
This code pulls in temperature data directly from xmacis (for MEFEC station), plots it as Tmax and Tmin on a 30-year rolling average by decade (1917-1946, 1927-1956, etc), and plots it either annually or by individual months.  Individual months are plotted separately or together as a grid for Tmax and for Tmin.  Individual Tmax and Tmin plots go directly to folders that can be opened and viewed in the base directory on a laptop.